WIP: Convert literals to docc links.
This implementation causes false-positive matches on the regex (see service.name notes). Really we should be matching against our known list of attribute IDs, and replace with calculated Docc symbol directly. That would however require that we pre-calculate Attribute Docc symbols, which changes the generator flow a bit. This work has yet to be done.